requirements.txt
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
redis==5.0.3

